vim: Revert back to two tabbing functions
I was wrong. Although I was always using `SetLocalTabbing`, what I
really want 99% of the time is to set tabbing across all buffers. I
discovered this when I realized that `Tab 2` in the .vimrc does very
little (it won't apply to any new buffers). So now the default for :Tab
is to set global tabbing behavior, and :TabLocal provides an escape
hatch.