vim: Allow hiding abandoned buffers
I'm experimenting with using tabs less, and that requires that I be able
to leave a buffer without saving it so that I can make changes across
files without having intermediate saves. This lets me keep the undo
history open for hidden buffers and continues to give me control over
when a buffer is saved.