bin: command to kill eslint
It runs as a daemon that gets screwed up sometimes (for unknown
reasons), but killing it, which forces a restart, seems to fix things.
When it's screwed up, it erroneously detects syntax errors in JS files,
blocking git commits because the pre-commit hook, which does a syntax
check on all modified JS files, fails.