ebook-curate: script to help edit ebook metadata
Helps automate editing ebook metadata. It makes it easier to fetch basic
metadata from the web, edit it before applying it to the book, set the
cover image from a URL, and move the book to the appropriate
subdirectory in our ebook library.

Depends on Calibre tools being available on the path. They're presently
stored at `~/.local/opt/ebook-tools/` with symlinks in `~/.local/bin/`.