vim: sticky background setting
I've been experimenting with different terminal colors recently, and
I've discovered to my dismay that vim on a server can't reliably detect
whether the terminal background is light or dark, so it always defaults
to light.

My solutions is to provide a vim function and command to easily change
the background color and remember the setting in a file so that we can
apply it every time vim starts up. This isn't ideal, but it turns out
that it's _quite difficult_ to propagate some local information (like
your terminal's background color) to a remote server over SSH.