git: show author time in log1
The prompt for this was that I'd reordered some commits all made on one
day, and when I pushed to GitHub, they were displayed in order by the
original author timestamps (not the commit order that I'd intended). In
fixing that situation with another rebase, it was helpful to be able to
see the log1 timeline I've grown used to along with their author time.