cr-leaders: Handle end of year
Since we're converting between timestamps and year-week strings in both
MySQL and PHP, we need to make sure that we're using the same convention
in both places. Failing to do so results in apparently-missing data for
weeks around the end of one year and the beginning of the next.

This also fixes a bug that we had in both places: failing to make the
year number match the convention for week numbers.