Add tat: a shortcut to attach to tmux sessions
The new `tat` is a shell function that uses tmux to lists sessions, fzf
to choose a session, and then (if a session was chosen), tmux again to
attach to the chosen session.

It has to be done via an alias or a shell function because a separate
script would happen in a sub-shell, and tmux invoked from a sub-shell
doesn't have a tty.

I chose a function so that I could introduce a local variable and avoid
polluting the global environment.