bash: work: Merge history between sessions
This is a little experimental, but the idea is to share bash history
between different tmux windows so that I don't need to know which window
I did something in to re-run an old command.

The downside is that it can feel kind of weird when going back up
through history because there are suddenly things there that didn't
happen in the current window. That's simultaneously the point and
unintuitive.