vim: set up PaperColor scheme
It's easy enough to make sure that it's installed with Vundle, and
hopefully a good color scheme can help me avoid unreadable text with,
e.g., spell mode.

I've recently realized that setting up a color scheme defaults the
background to light whether or not we're operating over SSH (at least on
Linux), so I'm moving the :BG command out of work config and into my
general vim config so that it's always available.

The PaperColor scheme has both a light and a dark variant, and I'm happy
enough with both (so far), so I'm using that for now. It also paints the
background, which _might_ be a little slow on a remote connection (we'll
have to see), but has the advantage that it always looks as intended,
whether the background color is "right" for the terminal or not.

Note: the PaperColor docs clued me in to the `yob` unimpaired mapping,
which will toggle the background color (the same as my :BG command, but
without persisting).