vim: add pencil
A plugin to make it more pleasant to edit markdown (or anything with long
lines) in vim.

With this addition, I should be much closer to comfortably editing my notes in
vim. Hey, I just noticed that pencil is already doing work for me in this
commit message by automatically hard-wrapping when I type in the middle of
a paragraph. Not bad, pencil.

For posterity: I learned that, in a terminal emulator, I can do :set columns=80
to soft wrap at 80 columns, even if the terminal is actually much wider.