rem: First draft of reminder script
This is a tool I've wanted for a long time. It's a CLI-driven reminder
application that uses `at` to queue up reminders and, when called without an
interactive terminal, uses `notify-send` to display notifications when
reminders are due, and uses `zenity` to display a list of pending reminders.
The goal is no mouse interaction and minimal typing in order to create
short-term reminders (usually no more than a few hours, but it _might_ work for
reminding oneself of something at a known time the next day). It

It's written entirely in Bash, but it's almost certainly not POSIX-compliant,
and it requires `at` to be set up, plus `libnotify-bin` for `notify-send`, plus
`zenity` if you're going to call it from the desktop environment (e.g., using
Fn-F2 in Gnome, or invoking it via a shortcut).