fasd: replace with z
I don't use FASD for anything besides switching directories, and it appears to
be unmaintained relative to `z`. Unfortunately, CtrlPZ only works with `fasd`,
but I'm okay with losing it because I haven't actually used it much in
practice.

Set up `zz`, a collaboration between `z` and `fzf` to filter `z` results with
`fzf`.