tmux: update style to match new color scheme
After changing my terminal's color scheme, I was having a hard time quickly
identifying which window was active. This does a better job (note that the
color names don't match the actual colors displayed). With my current color
scheme, the status style is a dark orange, the current window is green, and
inactive windows are blue.