bgrotate: use mtime to to "shuffle"
Instead of setting a seed based on time and shuffling images according to that,
update the mtime of displayed backgrounds and, on rotation, always choose the
image with the oldest mtime to display. To introduce some shuffling, swap the
mtime of the rotated-out image with that of a randomly-selected image from
a sliding window of most-recently-touched candidates. By tweaking the size of
the sliding window, we can trade off between more shuffling and longer
guaranteed intervals between seeing the same image twice.