bash: split out interactive config
I was thinking of bash_profile as the place to put config for interactive
shells, but that's wrong because many interactive shells are not login shells.
Identify interactive shells in one place and provide an easy hook to add
interactive configuration. Move the existing config that's only relevant in an
interactive context to the new bashrc.interactive.

This is all in preparation for adding GPG setup config, which requires
exporting GPG_TTY in interactive shells that are not login shells (i.e., in
shells spawned by a tmux session). Otherwise, new tmux windows don't get
GPG_TTY, which causes commit signing to fail.