rem: Add second resolution
`at` doesn't support sub-minute resolution, but we can fake it by resolving the
reminder time down to the second and passing that along in the command that
will run when the `at` timer (at minute resolution) expires; then that command
can compute the remaining number of seconds to the target reminder time, sleep
that many seconds if it's greater than 0, and _then_ trigger the notification.

Also, allow specifying an hour and minute as HHhMM (e.g., 10h14). I sometimes
use this notation elsewhere, and it's convenient to have it available here as
well.

Also, remove the unnecessary `function` keywords from the function definitions.