vim: increase textwidth to 99
It's the new standard at work, and I have enough screen real estate that
it makes sense as a default everywhere.

Configure git commit messages to stick to 72, using `set` instead of
`setlocal` to work around a bug in the pencil plugin, which I use for
markdown (and commit) editing.