Make FZF so much faster
OMG! I inadvertently mapped \fr in PHP buffers, and with the default
timeout between key sequences set to 1000ms, \f to open FZF was delaying
1s any time I was in a PHP buffer. I thought it was the system being
slow, but it was my own fault.

Let's remove that mapping because I use \f constantly and want it to be
fast. But let's also bring the timeout way down because I don't think
I need anywhere near a second.