bash: Prompt updates
Put the cursor in a consistent place on every prompt by moving to
a two-line layout, and since that means we care less how long the first
prompt line gets, add some more directory context, add the current git
branch (if any), and add the exit status of the last job if it was
non-zero. Also only show the count of background jobs if it's non-zero.