goto-browser: Switch browsers
I experimented with Vivaldi for a while, but ultimately found it too
slow. Additionally, although it had a bunch of features I really liked,
it had some weird behavior around keyboard shortcuts that got on my
nerves.

But having left Chrome behind for a while, it was a good opportunity to
try out a more vanilla Chromium. So far, it's going well.