pop-os-refresh: Try to capture my system conf
I did an OS refresh before I realized that my RAM was corrupt, which
clears everything but the home file system, so I had to rediscover all
of the apt-installed packages that I rely on day-to-day. Now I have
a place to write down my dependencies and semi-automate the process of
restoring.