db: A wrapper for mycli
To enable easily connecting to any of our databases from Cominor using
the mycli client. This is nice because mycli has some nice features for
changing table output style and auto-complete, and by using it from
Cominor (via a tunnel), it's easier to dump output to a file to process
further.