recents: Reword to be more flexible
I was curious about recently-modified files, and I remembered that I'd
written a recents command, but it turned out that it was very
specialized to listing recent markdown notes with a preview. I still
like that functionality, but I also want to be able to do things like
operate faster by skipping the preview, easily change the modification
time cutoff, include hidden files, and specify excludes. So I added some
option parsing and split the behavior into listing with a date stamp,
listing just paths, and listing both along with a preview.

I also decided to compromise on speed with the preview mode and first
find all matches, then display them so that I could align the previews
to the right of the paths without having to guess a reasonable fixed
path width.