fzg: Handle backticks in commit title
While integrating `fzg` behavior into `gh-commits`, I realized that we
were executing text in backticks as part of the preview. That's
dangerous! Honestly, I still don't fully understand why because
I thought that the line coming out of fzf is supposed to be
single-quoted, but I learned that I can use fzf's field index
expressions to have fzf pull out the commit SHA instead of going through
a pipe to `cut`. That circumvents putting the whole log line into the
pipeline, so we're safe.

I also did a little minor refactoring to rename some things that I found
confusing on a second visit and cut an unnecessary helper function.