Vim: Switch to monokai-soda theme
I happened to see it in bat, I liked it, and it meets my requirements of
having both a kitty theme and a vim theme, neither of which messes with
the native background color.

I had to fork the vim theme to update the implementation to the latest
version of colorbuddy.