wcl: Add world clock
I've been playing with this off-and-on to make a simple world clock at
the command line. The primary features are that it highlights your
current timezone, highlights timezones that fall on the previous or next
day (relative to the current timezone), displays timezone offsets both
relative to UTC and the local timezone, and that it allows you to
specify a date-time.