toggle-headset-profile: Add GUI
When I'm not playing music, it isn't easy to know which mode I'm in. Now
we can use zenity to display a GUI with the currently-selected profile
and the ability to change it. I also (I think) improved some of the
error handling such that the script does a better job of exiting with an
error status when something doesn't work as expected.